feat: add initial unit tests with Vitest for core components and reducer
Well, we finally did it. We added tests. Unit tests, to be precise. Using Vitest, because I could find a good jest.

Changes include:
- Vitest: The new sheriff in town.
- Initial unit tests for:
  - FeatureCard: Because cards need love too.
  - CategoryItems: Sorting out categories, one test at a time.
  - MainContent: For when your content needs to be main.
  - Home: There's no place like home, especially when it's bug-free.
- Reducer test: To make sure our state management isn’t reducing our sanity.
- Mocked preact-router: Because why not make routes do what we want?

All tests are passing, which means either everything works perfectly or our tests are as good as those expired milk cartons in the fridge. Either way, we’re off to a good start.